Moved Adaptor (no idea where to put it honestly) and then added the card view.
