+++ app/src/main/java/com/josh_petro/hudlu/MyAdapter.java
@@ -1,12 +1,10 @@
-package com.wasinaso.joshpetro.application;
+package com.josh_petro.hudlu;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
-
-import com.josh_petro.hudlu.R;
 
 /**
  * Created by joshpetro on 3/28/16.
