+++ app/src/main/res/layout/activity_main.xml
@@ -15,7 +15,24 @@
         android:scrollbars="vertical"
         android:layout_width="match_parent"
         android:layout_height="match_parent"/>
+
     <TextView android:text="Hello Hudl!" android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:textColor="#ffffff" />
+
+    <!-- A CardView that contains a TextView -->
+    <android.support.v7.widget.CardView
+        xmlns:card_view="http://schemas.android.com/apk/res-auto"
+        android:id="@+id/my_card_view"
+        android:layout_gravity="center"
+        android:layout_width="200dp"
+        android:layout_height="200dp"
+        card_view:cardCornerRadius="4dp">
+
+        <TextView
+            android:id="@+id/info_text"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:text="TEST"/>
+    </android.support.v7.widget.CardView>
 </RelativeLayout>
